requirements.txt
tensorflow-cpu
tensorflow_hub